Refactor views into Activites and Fragment packages
