+++ app/src/main/res/layout/fragment_bookmark.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Views.Bookmarks">
+    tools:context=".Views.Fragment.Bookmarks">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
